requirements.txt
transformers==4.48.0
Pillow
torch
torchvision
numpy
